requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements/dev.txt requirements/base.in requirements/test-master.in requirements/dev.in requirements/quality.in requirements/reporting.in
#
amqp==1.4.9               # via kombu
anyjson==0.3.3            # via kombu
argparse==1.4.0           # via caniusepython3
astroid==1.5.2            # via edx-lint, pylint, pylint-celery, pylint-plugin-utils
backports.functools-lru-cache==1.4  # via astroid, caniusepython3, pylint
billiard==3.3.0.23        # via celery
caniusepython3==5.0.0
celery==3.1.18
click-log==0.1.8          # via edx-lint
click==6.7                # via click-log, edx-lint, pip-tools
configparser==3.5.0       # via pylint
diff-cover==0.9.12
distlib==0.2.5            # via caniusepython3
django-config-models==0.1.8
django-extensions==1.5.9
django-filter==1.0.4
django-model-utils==2.3.1
django-object-actions==0.10.0
django-simple-history==1.6.3
django-waffle==0.12.0
django==1.8.18
djangorestframework-jwt==1.11.0  # via edx-drf-extensions
djangorestframework-oauth==1.1.0
djangorestframework==3.6.3
edx-django-oauth2-provider==1.1.4
edx-drf-extensions==1.2.3
edx-i18n-tools==0.3.10
edx-lint==0.5.4
edx-opaque-keys==0.4.0
edx-rest-api-client==1.7.1
enum34==1.1.6             # via astroid
first==2.0.1              # via pip-tools
flaky==3.3.0
future==0.16.0            # via vertica-python
futures==3.1.1            # via caniusepython3
inflect==0.2.5            # via jinja2-pluralize
isort==4.2.15
jinja2-pluralize==0.3.0   # via diff-cover
jinja2==2.9.6             # via diff-cover, jinja2-pluralize
jsonfield==2.0.2
kombu==3.0.37             # via celery
lazy-object-proxy==1.3.1  # via astroid
markupsafe==1.0           # via jinja2
mccabe==0.6.1             # via pylint
packaging==16.8           # via caniusepython3
path.py==10.4             # via edx-i18n-tools
pbr==3.1.1                # via stevedore
pillow==3.4
pip-tools==1.9.0
pkginfo==1.4.1            # via twine
pluggy==0.5.2             # via tox
polib==1.0.8              # via edx-i18n-tools
py==1.4.34                # via tox
pycodestyle==2.3.1
pydocstyle==2.0.0
pygments==2.2.0           # via diff-cover
pyjwt==1.5.3              # via djangorestframework-jwt, edx-rest-api-client
pylint-celery==0.3        # via edx-lint
pylint-django==0.7.2      # via edx-lint
pylint-plugin-utils==0.2.6  # via pylint-celery, pylint-django
pylint==1.7.1             # via edx-lint, pylint-celery, pylint-django, pylint-plugin-utils
pyminizip==0.2.1
pymongo==3.5.1            # via edx-opaque-keys
pyparsing==2.2.0          # via packaging
python-dateutil==2.6.1    # via edx-drf-extensions, vertica-python
pytz==2017.2              # via celery, vertica-python
pyyaml==3.12              # via edx-i18n-tools
requests-toolbelt==0.8.0  # via twine
requests==2.9.1
shortuuid==0.5.0          # via edx-django-oauth2-provider
singledispatch==3.4.0.3   # via astroid, pylint
six==1.11.0               # via astroid, diff-cover, django-extensions, edx-i18n-tools, edx-lint, edx-opaque-keys, packaging, pip-tools, pydocstyle, pylint, python-dateutil, singledispatch, stevedore, vertica-python
slumber==0.7.1            # via edx-rest-api-client
snowballstemmer==1.2.1    # via pydocstyle
stevedore==1.27.0         # via edx-opaque-keys
testfixtures==5.2.0
tox-battery==0.2.0
tox==2.8.2
tqdm==4.17.1              # via twine
twine==1.9.1
unicodecsv==0.14.1
vertica-python==0.7.3
virtualenv==15.1.0        # via tox
wheel==0.30.0
wrapt==1.10.11            # via astroid
